% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{directRandomKsCalc}
\alias{directRandomKsCalc}
\title{#' @export
memoRandomV = memoise::memoise(randomV)}
\usage{
directRandomKsCalc(length, n, d)
}
\description{
#' @export
memoRandomV = memoise::memoise(randomV)
}
