% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{MSigDB62}
\alias{MSigDB62}
\title{MsigDB 6.2 tags}
\format{An object of class \code{list} of length 3636.}
\usage{
MSigDB62
}
\description{
Up and down tags from the version 6.2 of MsigDB. Used for calculation
of specificity and reliability ratings
}
\keyword{datasets}
