% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{MSigDBLegacy}
\alias{MSigDBLegacy}
\title{Legacy MsigDB tags}
\format{An object of class \code{list} of length 312.}
\usage{
MSigDBLegacy
}
\description{
Legacy MsigDB tags used in the Lamb et al 2006 paper. Used for calculation
of specificity and reliability ratings
}
\keyword{datasets}
