% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{score}
\alias{score}
\title{#' @export
memoDirectRandomKsCalc = memoise::memoise(directRandomKsCalc)}
\usage{
score(kUp, kDown)
}
\description{
#' @export
memoDirectRandomKsCalc = memoise::memoise(directRandomKsCalc)
}
